Move the rigid body calculations to separate header, still needs some cleanup but works!
