+++ imgui.ini
@@ -9,7 +9,7 @@ Size=476,528
 Collapsed=1
 
 [Window][Editor]
-Pos=916,95
+Pos=918,96
 Size=408,604
 Collapsed=0
